Split the platform surface out of the IPC monolith
The cross-cutting services rather than a vault type: MCP, Nodus Server, the Word and
Zotero add-ins, app info, AI providers and local runtimes, images, audio and TTS,
translations, the Zotero client, the academic home snapshot, shell links, and the
three API-key channels.

  electron/ipc.ts     1187-1469 -> electron/ipc/platform.ts     (88 handlers)
  electron/preload.ts   242-397 -> electron/preload/platform.ts (98 bindings)
  shared/types.ts     7435-7582 -> shared/api/platform.ts       (94 signatures)

A corpus block — the works: and ideas: handlers, bindings and signatures — sat inside
the platform range in all three files and was relocated out first, one step per file.
extensionForOriginalImage and originalImageFileName, module-level helpers used only by
images:, moved into the slice with it.

A first attempt at this domain was reverted rather than repaired, and the reason is
worth recording because it was misdiagnosed at the time. It looked like the remaining
domains shared module-level helpers across boundaries, which would have been a real
architectural blocker. They do not: tracing every call site shows withVaultKeyProviders,
vaultBusyMessage and vaultSwitchMessage are used only by vaults:, processFullChain only
by works:, MANUAL_IDEA_MARKER only by notes:. The actual cause was a helper-relocation
script that cut from the nearest preceding /** — which was 70 lines above its target —
and dragged four unrelated functions into the new module, after which an
ESLint-driven import pruner ran on the damaged file and removed valid imports. The
relocation is now line-exact and the pruner is no longer used on this path.

Where the split stands: 682 of 1,206 channels — a clear majority — live in a domain
module, 524 remain in electron/ipc.ts. That file is 2,451 lines (from 6,013 — down
59%), electron/preload.ts 916 (from 1,862 — down 51%), shared/types.ts 8,220.

Six more tests moved onto the census sentinels, including several that resolve paths
against a local `root` rather than `repoRoot`.

Verified: 1,425 tests, lint and typecheck clean, contract test green.
