+++ electron/ipc.ts
@@ -3,19 +3,14 @@ import fs from 'node:fs';
 import os from 'node:os';
 import AdmZip from 'adm-zip';
 import { ipcMain, shell, BrowserWindow, dialog, app } from 'electron';
-import { originalImagePayloadFromUrl } from './imageProtocol';
 import { seedTestimonyDemoData } from './db/testimonyDemoData';
 
 import {
-  openPrivacyPolicy,
   showImportOpenDialog,
 } from './privacy';
 import type {
   AppLanguage,
   AppSettings,
-  AudioEntityKind,
-  AudioProvider,
-  AudioSegmentRequest,
   AddProjectLinkInput,
   ApplyManuscriptCitationRequest,
   ApplyProjectSuggestionsRequest,
@@ -23,14 +18,10 @@ import type {
   ChapterSuggestionStatus,
   QueueKind,
   WorkFilter,
-  AiProvider,
-  LocalProvider,
   ModelRef,
   ZoteroItem,
-  ZoteroLibrary,
   ResearchChatRequest,
   ReadingPathRequest,
-  EmbeddingProvider,
   UpdateCheckResponse,
   ChatMessageRecord,
   ResearchContextSelection,
@@ -44,8 +35,6 @@ import type {
   WritingWorkshopDraftRequest,
   WritingWorkshopExportRequest,
   WritingWorkshopSaveDraftRequest,
-  TranslationEntityKind,
-  GenerateTranslationRequest,
   DeepResearchRequest,
   DebateAnalysisRequest,
   RqDecomposeRequest,
@@ -95,9 +84,6 @@ import type {
   RelationshipSubtype,
   SocialContactInput,
   SocialRelationInput,
-  DecorativeImageActionRequest,
-  DecorativeImageEntityKind,
-  DecorativeImageStyle,
   SearchResultKind,
   CreateStudyAcademicYearInput,
   UpdateStudyAcademicYearInput,
@@ -147,7 +133,6 @@ import type {
   StudyAttemptStartInput,
   StudyTestBuildRequest,
   StudyRubricInput,
-  StudyPronunciationEntry,
 } from '@shared/types';
 
 // Mirrors MANUAL_IDEA_MARKER in shared/types.ts. Defined locally because the
@@ -165,29 +150,20 @@ import { registerDatabasesIpc } from './ipc/databases';
 import { registerPrimarySourcesIpc } from './ipc/primarySources';
 import { registerArchiveIpc } from './ipc/archive';
 import { registerWorldbuildingIpc } from './ipc/worldbuilding';
+import { registerPlatformIpc } from './ipc/platform';
 import {
-  connectMcpTunnel,
-  disconnectMcpTunnel,
-  forgetMcpTunnel,
-  getMcpStatus,
-  getMcpTunnelStatus,
-  regenerateMcpToken,
   restartMcpServer,
-  restartMcpTunnelIfConfigured,
   startMcpServer,
   startMcpTunnelIfConfigured,
   stopMcpServer,
   stopMcpTunnel,
 } from './mcp';
-import { getCopilotStatus, regenerateCopilotToken, restartCopilotServer, startCopilotServer, stopCopilotServer } from './copilot/server';
+import { restartCopilotServer, startCopilotServer, stopCopilotServer } from './copilot/server';
 import {
-  getZoteroPluginStatus,
-  regenerateZoteroPluginToken,
   restartZoteroPluginServer,
   startZoteroPluginServer,
   stopZoteroPluginServer,
 } from './zotero-plugin/server';
-import { exportZoteroPluginXpi, getZoteroInstallInfo, installZoteroPlugin } from './zotero-plugin/install';
 import {
   applyMascotWindow,
   beginMascotWindowDrag,
@@ -207,65 +183,22 @@ import { getNodiViewContext, setNodiViewContext, streamNodiChat } from './ai/nod
 import type { NodiChatRequest } from '@shared/types';
 import { clearNodiConversations, deleteNodiConversation, getNodiConversation, listNodiConversations, saveNodiConversation } from './nodiConversations';
 import { deleteNodiNote, listNodiNotes, saveNodiNote } from './nodiNotes';
-import { ensureCopilotCert } from './copilot/certs';
-import { installCopilotAddin, installLibreOfficeCopilot } from './copilot/install';
-import { setApiKey, clearApiKey, getApiKey, copyApiKeysBetweenVaults, listApiKeyProvidersForVault, setBackupPassword, clearBackupPassword, hasBackupPassword, getBackupPassword, getBackupRecoveryKey } from './secrets/secretStore';
-import { recoverLegacyApiKeys } from './secrets/legacySecretRecovery';
+import { copyApiKeysBetweenVaults, listApiKeyProvidersForVault, setBackupPassword, clearBackupPassword, hasBackupPassword, getBackupPassword, getBackupRecoveryKey } from './secrets/secretStore';
 import { runAutoBackupNow } from './export/autoBackup';
 import { MIN_BACKUP_PASSWORD_LENGTH } from './export/backupCrypto';
-import { listEmbeddingModels, listModels, testLocalProvider } from './ai/providers';
 import {
-  cancelChatGptSubscriptionLogin,
-  getChatGptSubscriptionStatus,
-  listChatGptSubscriptionModels,
-  logoutChatGptSubscription,
   onChatGptSubscriptionStatusChanged,
-  startChatGptSubscriptionLogin,
 } from './ai/codexSubscription';
 import {
-  cancelGitHubCopilotSubscriptionLogin,
-  getGitHubCopilotSubscriptionStatus,
-  listGitHubCopilotSubscriptionModels,
-  logoutGitHubCopilotSubscription,
   onGitHubCopilotSubscriptionStatusChanged,
-  startGitHubCopilotSubscriptionLogin,
 } from './ai/githubCopilotSubscription';
-import { getOpenCodeGoUsageStatus, onOpenCodeGoUsageStatusChanged } from './ai/openCodeGoUsage';
-import { listImageModels } from './ai/imageModels';
+import { onOpenCodeGoUsageStatusChanged } from './ai/openCodeGoUsage';
 import {
   applyDecorativeImageOption,
-  deleteDecorativeImage,
   generatePersonPortraitFromDescription,
   interruptDecorativeImageGenerations,
   invalidateDecorativeImageGeneration,
-  queueDecorativeImageGeneration,
-  revertDecorativeImage,
-  saveCustomDecorativeImage,
 } from './ai/decorativeImages';
-import { getDecorativeImage, getDecorativeImageData } from './db/decorativeImagesRepo';
-import {
-  clearEntityClips,
-  deleteClip as deleteAudioClip,
-  deleteEntityClips,
-  getEntitySegments,
-  audioClipPath,
-  createStudyAudioBookmark,
-  deleteStudyAudioBookmark,
-  getStudyPronunciations,
-  listStudyAudioBookmarks,
-  listStudyAudioPlaylist,
-  listEntityClips,
-  readClipBytes,
-  saveClip,
-  setStudyPronunciations,
-} from './audio/audioService';
-import {
-  clearHumeKey,
-  humeHasKey,
-  listHumeVoices,
-  setHumeKey,
-  synthesizeHume,
-} from './audio/hume';
 import * as zotero from './zotero/zoteroClient';
 import * as works from './db/worksRepo';
 import { reconcileAuthorLayerOnce } from './db/authorsRepo';
@@ -279,14 +212,9 @@ import { aggregateGaps, aggregateGapsPage, contradictionCount, getGapDetail } fr
 import { getSyncLog } from './db/syncRepo';
 import { fullSync, ingestZoteroItem, startRealtimeSync, stopRealtimeSync } from './sync/syncService';
 import {
-  disconnectNodusServerVault,
-  getNodusServerOverview,
-  pairNodusServer,
   restartNodusServerSync,
-  setNodusServerLanguage,
   startNodusServerSync,
   stopNodusServerSync,
-  syncNodusServerVaultNow,
 } from './serverSync/serverSyncService';
 import { scanQueue } from './pipeline/scanQueue';
 import { buildIdeaGraph, buildIdeaGraphOverview, buildIdeaThemeGraph, buildAuthorGraph, getContradictions, getDebates, buildReadingPath } from './graph/graphService';
@@ -328,7 +256,6 @@ import { getSearchResultDetail, globalSearch } from './db/searchRepo';
 import { semanticSearch, findSimilarToIdea } from './ai/semanticSearch';
 import { listSavedSearches, saveSearch, deleteSavedSearch } from './db/savedSearchesRepo';
 import { getCorpusHealth } from './db/corpusHealthRepo';
-import { getAcademicHomeStats } from './db/homeRepo';
 import { analyzeChapterRelations, getChapterRelations, onChapterRelationsProgress } from './ai/chapterIdeas';
 import { applyManuscriptCitation, verifyManuscriptCitations } from './ai/manuscriptVerifier';
 import { suggestGapSearch } from './ai/gapSearch';
@@ -363,19 +290,6 @@ import {
   transcribeWhisperCpp,
   installWhisperCpp, uninstallWhisperCpp,
 } from './stt/whisperCpp';
-import {
-  cancelNodusLocalDownloads,
-  deleteNodusLocalModel,
-  downloadNodusLocalModel,
-  getNodusLocalAiStatus,
-  installNodusLocalRuntime,
-} from './ai/nodusLocalAi';
-import {
-  deleteNodusLocalImageModel,
-  downloadNodusLocalImageModel,
-  getNodusLocalImageStatus,
-  installNodusLocalImageRuntime,
-} from './ai/nodusLocalImages';
 import { improveStudyText } from './ai/studyImprove';
 import * as studySearch from './ai/studySearch';
 import * as studyAssistant from './ai/studyAssistant';
@@ -423,8 +337,6 @@ import * as manualIdeas from './db/manualIdeasRepo';
 import * as tutorRoutes from './db/tutorRepo';
 import * as writingDrafts from './db/writingDraftsRepo';
 import * as translationsRepo from './db/translationsRepo';
-import { TRANSLATION_LANGUAGES } from '@shared/types';
-import { translateMarkdown, titleFromMarkdown } from './ai/translate';
 import * as workSummaries from './db/workSummariesRepo';
 import * as projects from './db/projectsRepo';
 import { closeDb, getDb } from './db/database';
@@ -599,29 +511,7 @@ async function importStudyMaterialPaths(paths: string[], input: StudyMaterialImp
 }
 
 
-function extensionForOriginalImage(mime: string): string {
-  switch (mime.toLowerCase()) {
-    case 'image/png': return 'png';
-    case 'image/webp': return 'webp';
-    case 'image/gif': return 'gif';
-    case 'image/bmp': return 'bmp';
-    case 'image/tiff': return 'tiff';
-    case 'image/svg+xml': return 'svg';
-    case 'image/avif': return 'avif';
-    default: return 'jpg';
-  }
-}
 
-function originalImageFileName(label: string | null | undefined, mime: string): string {
-  const base =
-    (label ?? '')
-      .replace(/\.[a-z0-9]{2,5}$/i, '')
-      .replace(/[^\p{L}\p{N} _-]/gu, '')
-      .trim()
-      .slice(0, 120)
-    || 'imagen';
-  return `${base}.${extensionForOriginalImage(mime)}`;
-}
 
 /** Register every IPC channel backing the window.nodus API. */
 export function registerIpc(
@@ -635,6 +525,7 @@ export function registerIpc(
   // Domains extracted from this file, each owning its own channels and repo
   // imports. What remains below is everything not yet split out.
   registerProsopographyIpc(context);
+  registerPlatformIpc(context);
   registerWorldbuildingIpc(context);
   registerArchiveIpc(context);
   registerPrimarySourcesIpc(context);
@@ -1184,274 +1075,8 @@ export function registerIpc(
 
 
 
-  h('mcp:status', async () => getMcpStatus());
-  h('mcp:regenerateToken', async () => {
-    await stopMcpTunnel();
-    const token = await regenerateMcpToken();
-    void restartMcpTunnelIfConfigured();
-    return token;
-  });
-  h('mcp:tunnel:status', async () => getMcpTunnelStatus());
-  h('mcp:tunnel:connect', async (_e, input) => connectMcpTunnel(input));
-  h('mcp:tunnel:disconnect', async () => disconnectMcpTunnel());
-  h('mcp:tunnel:forget', async () => forgetMcpTunnel());
-  h('nodusServer:overview', async () => getNodusServerOverview());
-  h('nodusServer:pair', async (_e, url: string, code: string) => pairNodusServer(url, code));
-  h('nodusServer:setLanguage', async (_e, language: AppLanguage, vaultId?: string) => setNodusServerLanguage(language, vaultId));
-  h('nodusServer:syncVaultNow', async (_e, vaultId: string) => syncNodusServerVaultNow(vaultId));
-  h('nodusServer:disconnectVault', async (_e, vaultId: string) => disconnectNodusServerVault(vaultId));
-  h('copilot:status', async () => getCopilotStatus());
-  h('copilot:regenerateToken', async () => regenerateCopilotToken());
-  h('copilot:ensureCert', async () => {
-    const result = await ensureCopilotCert(getSettings().uiLanguage);
-    if (result.ok && getSettings().copilotEnabled) await restartCopilotServer();
-    return result;
-  });
-  h('copilot:installAddin', async () => installCopilotAddin(app.getAppPath(), app.getVersion()));
-  h('copilot:installLibreOffice', async () => installLibreOfficeCopilot(app.getAppPath()));
-  h('zoteroPlugin:status', async () => getZoteroPluginStatus());
-  h('zoteroPlugin:regenerateToken', async () => regenerateZoteroPluginToken());
-  h('zoteroPlugin:installInfo', async () => getZoteroInstallInfo());
-  h('zoteroPlugin:install', async () => installZoteroPlugin());
-  h('zoteroPlugin:downloadXpi', async () => exportZoteroPluginXpi());
-  h('app:info', async () => {
-    const osName =
-      process.platform === 'darwin' ? 'macOS' : process.platform === 'win32' ? 'Windows' : process.platform === 'linux' ? 'Linux' : process.platform;
-    return {
-      version: app.getVersion(),
-      platform: process.platform,
-      osName,
-      osVersion: os.release(),
-      arch: process.arch,
-      electron: process.versions.electron ?? '',
-    };
-  });
-  h('settings:setApiKey', async (_e, provider: AiProvider, key: string) => setApiKey(provider, key));
-  h('settings:clearApiKey', async (_e, provider: AiProvider) => clearApiKey(provider));
-  h('settings:recoverApiKeys', async (event) => {
-    const result = await recoverLegacyApiKeys();
-    if (result.recoveredProviders.length > 0) {
-      const settings = getSettings();
-      if (settings.autoBackupEnabled && settings.autoBackupFolder) await runAutoBackupNow(app.getVersion());
-    }
-    event.sender.send('settings:apiKeysRecovered', result);
-    return result;
-  });
 
-  h('ai:chatgptSubscription:status', async () => getChatGptSubscriptionStatus());
-  h('ai:chatgptSubscription:login', async () => startChatGptSubscriptionLogin());
-  h('ai:chatgptSubscription:cancelLogin', async (_event, loginId: string) =>
-    cancelChatGptSubscriptionLogin(loginId)
-  );
-  h('ai:chatgptSubscription:logout', async () => logoutChatGptSubscription());
-  h('ai:githubCopilotSubscription:status', async () => getGitHubCopilotSubscriptionStatus());
-  h('ai:githubCopilotSubscription:login', async () => startGitHubCopilotSubscriptionLogin());
-  h('ai:githubCopilotSubscription:cancelLogin', async () => cancelGitHubCopilotSubscriptionLogin());
-  h('ai:githubCopilotSubscription:logout', async () => logoutGitHubCopilotSubscription());
-  h('ai:openCodeGo:usage', async () => getOpenCodeGoUsageStatus());
-
-  // AI model discovery (OpenRouter needs no key; others use the stored key).
-  h('ai:listModels', async (_e, provider: AiProvider) => {
-    if (provider === 'codex') return listChatGptSubscriptionModels();
-    if (provider === 'github-copilot') return listGitHubCopilotSubscriptionModels();
-    return listModels(provider, getApiKey(provider));
-  });
-  h('ai:listEmbeddingModels', async (_e, provider: EmbeddingProvider) =>
-    listEmbeddingModels(provider, getApiKey(provider))
-  );
-  h('ai:testLocalProvider', async (_e, provider: LocalProvider) => testLocalProvider(provider, getApiKey(provider)));
-  h('ai:listImageModels', async () => listImageModels());
-  h('ai:nodusLocal:status', async () => getNodusLocalAiStatus());
-  h('ai:nodusLocal:installRuntime', async (event, requestId: string) =>
-    installNodusLocalRuntime((fraction) => {
-      if (!event.sender.isDestroyed()) event.sender.send('ai:nodusLocal:progress', requestId, fraction);
-    }));
-  h('ai:nodusLocal:downloadModel', async (event, requestId: string, model: string) =>
-    downloadNodusLocalModel(model, (fraction) => {
-      if (!event.sender.isDestroyed()) event.sender.send('ai:nodusLocal:progress', requestId, fraction);
-    }));
-  h('ai:nodusLocal:cancelDownloads', async () => cancelNodusLocalDownloads());
-  h('ai:nodusLocal:deleteModel', async (_event, model: string) => deleteNodusLocalModel(model));
-  h('ai:nodusLocalImage:status', async () => getNodusLocalImageStatus());
-  h('ai:nodusLocalImage:installRuntime', async (event, requestId: string) =>
-    installNodusLocalImageRuntime((fraction) => {
-      if (!event.sender.isDestroyed()) event.sender.send('ai:nodusLocalImage:progress', requestId, fraction);
-    }));
-  h('ai:nodusLocalImage:downloadModel', async (event, requestId: string, model: string) =>
-    downloadNodusLocalImageModel(model, (fraction) => {
-      if (!event.sender.isDestroyed()) event.sender.send('ai:nodusLocalImage:progress', requestId, fraction);
-    }));
-  h('ai:nodusLocalImage:deleteModel', async (_event, model: string) => deleteNodusLocalImageModel(model));
-  h('images:get', async (_e, entityKind: DecorativeImageEntityKind, entityId: string) =>
-    getDecorativeImage(entityKind, entityId)
-  );
-  h('images:data', async (_e, entityKind: DecorativeImageEntityKind, entityId: string, thumbnail?: boolean) => {
-    const data = getDecorativeImageData(entityKind, entityId, thumbnail);
-    return data ? `data:${data.mimeType};base64,${data.bytes.toString('base64')}` : null;
-  });
-  h('images:queue', async (e, request: DecorativeImageActionRequest) =>
-    queueDecorativeImageGeneration(request, (image) => {
-      if (!e.sender.isDestroyed()) e.sender.send('images:changed', image);
-    })
-  );
-  h('images:upload', async (
-    e,
-    entityKind: DecorativeImageEntityKind,
-    entityId: string,
-    bytes: Uint8Array,
-    mimeType?: string,
-    style?: DecorativeImageStyle
-  ) => {
-    const image = await saveCustomDecorativeImage(entityKind, entityId, Buffer.from(bytes), mimeType, style);
-    if (!e.sender.isDestroyed()) e.sender.send('images:changed', image);
-    return image;
-  });
-  h('images:revert', async (e, entityKind: DecorativeImageEntityKind, entityId: string) => {
-    const image = revertDecorativeImage(entityKind, entityId);
-    if (!e.sender.isDestroyed()) e.sender.send('images:changed', image);
-    return image;
-  });
-  h('images:delete', async (_e, entityKind: DecorativeImageEntityKind, entityId: string) =>
-    deleteDecorativeImage(entityKind, entityId)
-  );
-  h('images:downloadOriginal', async (_e, source: string, label?: string | null) => {
-    const image = originalImagePayloadFromUrl(source);
-    if (!image) throw new Error('No se encontró la imagen original.');
-    const picked = await dialog.showSaveDialog(getWindow() ?? undefined!, {
-      title: 'Descargar imagen original',
-      defaultPath: originalImageFileName(label, image.mime),
-      filters: [{
-        name: 'Imagen original',
-        extensions: [extensionForOriginalImage(image.mime)],
-      }],
-    });
-    if (picked.canceled || !picked.filePath) return { canceled: true, path: null };
-    fs.writeFileSync(picked.filePath, image.blob);
-    return { canceled: false, path: picked.filePath };
-  });
-
-  // audio / text-to-speech. Synthesis runs in the renderer (Piper via WebAssembly);
-  // the main process supplies the speakable segments and persists the resulting WAVs.
-  h('audio:segments', async (_e, entityKind: AudioEntityKind, entityId: string, request?: AudioSegmentRequest) =>
-    getEntitySegments(entityKind, entityId, request)
-  );
-  h('audio:listClips', async (_e, entityKind: AudioEntityKind, entityId: string) =>
-    listEntityClips(entityKind, entityId)
-  );
-  h('audio:clearClips', async (_e, entityKind: AudioEntityKind, entityId: string) => {
-    clearEntityClips(entityKind, entityId);
-  });
-  h('audio:saveClip', async (
-    _e,
-    entityKind: AudioEntityKind,
-    entityId: string,
-    input: { segmentIndex: number; segmentLabel: string; provider: AudioProvider; voice: string; language: string; bytes: Uint8Array }
-  ) => saveClip(entityKind, entityId, { ...input, bytes: input.bytes }));
-  h('audio:clipData', async (_e, clipId: string) => {
-    const data = readClipBytes(clipId);
-    return data ? `data:${data.mime};base64,${data.bytes.toString('base64')}` : null;
-  });
-  h('audio:deleteClip', async (_e, clipId: string) => {
-    deleteAudioClip(clipId);
-  });
-  h('audio:deleteEntityClips', async (_e, entityKind: AudioEntityKind, entityId: string) => {
-    deleteEntityClips(entityKind, entityId);
-  });
-  h('audio:exportClip', async (_e, clipId: string) => {
-    const source = audioClipPath(clipId); if (!source) return null;
-    const picked = await dialog.showSaveDialog(getWindow() ?? undefined!, { title: 'Guardar audio', defaultPath: path.basename(source), filters: [{ name: 'Audio WAV', extensions: ['wav'] }] });
-    if (picked.canceled || !picked.filePath) return null; fs.copyFileSync(source, picked.filePath); return { path: picked.filePath };
-  });
-  h('audio:study:bookmarks', async (_e, kind: AudioEntityKind, id: string) => listStudyAudioBookmarks(kind, id));
-  h('audio:study:bookmark:create', async (_e, kind: AudioEntityKind, id: string, segmentIndex: number, label: string) => createStudyAudioBookmark(kind, id, segmentIndex, label));
-  h('audio:study:bookmark:delete', async (_e, id: string) => deleteStudyAudioBookmark(id));
-  h('audio:study:pronunciations', async (_e, subjectId: string) => getStudyPronunciations(subjectId));
-  h('audio:study:pronunciations:set', async (_e, subjectId: string, entries: StudyPronunciationEntry[]) => setStudyPronunciations(subjectId, entries));
-  h('audio:study:playlist', async (_e, subjectId: string) => listStudyAudioPlaylist(subjectId));
-
-  // Hume (cloud TTS): key stays in the main process; the renderer only sees
-  // whether a key exists, the voice list, and the resulting audio bytes.
-  h('audio:humeStatus', async () => ({ hasKey: humeHasKey() }));
-  h('audio:humeSetKey', async (_e, key: string) => {
-    setHumeKey(key);
-    return { hasKey: humeHasKey() };
-  });
-  h('audio:humeClearKey', async () => {
-    clearHumeKey();
-    return { hasKey: humeHasKey() };
-  });
-  h('audio:humeVoices', async (_e, language?: string) => listHumeVoices(language));
-  h('audio:humeSynthesize', async (_e, voiceId: string, provider: 'HUME_AI' | 'CUSTOM_VOICE', text: string) => {
-    const bytes = await synthesizeHume(voiceId, provider, text);
-    return new Uint8Array(bytes);
-  });
-
-  // AI translations. The renderer assembles an entity's Markdown and passes it in;
-  // the main process translates it (chunked, preserving citations) and stores one
-  // copy per language.
-  h('translations:list', async (_e, entityKind: TranslationEntityKind, entityId: string) =>
-    translationsRepo.listContentTranslations(entityKind, entityId)
-  );
-  h('translations:get', async (_e, id: string) => translationsRepo.getContentTranslation(id));
-  h('translations:generate', async (_e, request: GenerateTranslationRequest) => {
-    const language = TRANSLATION_LANGUAGES.find((l) => l.code === request.language);
-    if (!language) throw new Error(`Idioma de traducción no soportado: ${request.language}`);
-    const source = request.sourceMarkdown.trim();
-    if (!source) throw new Error('No hay contenido para traducir.');
-    const pending = translationsRepo.beginContentTranslation({ entityKind: request.entityKind, entityId: request.entityId, language: language.code, languageLabel: language.nativeName, sourceTitle: request.sourceTitle, model: request.model ?? null });
-    try {
-      const markdown = await translateMarkdown({ markdown: source, language, model: request.model });
-      return translationsRepo.upsertContentTranslation({
-        entityKind: request.entityKind, entityId: request.entityId, language: language.code, languageLabel: language.nativeName,
-        title: titleFromMarkdown(markdown, request.sourceTitle), markdown, model: request.model ?? null,
-      });
-    } catch (cause) {
-      translationsRepo.failContentTranslation(pending.id, cause instanceof Error ? cause.message : String(cause));
-      throw cause;
-    }
-  });
-  h('translations:delete', async (_e, id: string) => {
-    translationsRepo.deleteContentTranslation(id);
-  });
-
-  // zotero
-  h('zotero:ping', async () => {
-    const res = await zotero.ping();
-    // Local API always uses users/0; persist that so all later calls address it correctly.
-    if (res.ok) updateSettings({ zoteroUserId: zotero.LOCAL_USER_ID });
-    return res;
-  });
-  h('zotero:libraries', async () => zotero.libraries());
-  h('zotero:collections', async (_e, library?: ZoteroLibrary) => {
-    const { zoteroUserId } = getSettings();
-    return zotero.topCollections(zoteroUserId, library);
-  });
-  h('zotero:childCollections', async (_e, parentKey: string, library?: ZoteroLibrary) => {
-    const { zoteroUserId } = getSettings();
-    return zotero.childCollections(zoteroUserId, parentKey, library);
-  });
-  h('zotero:collectionItems', async (_e, collectionKey: string, opts?: { query?: string; recursive?: boolean; library?: ZoteroLibrary }) => {
-    const { zoteroUserId } = getSettings();
-    return opts?.recursive
-      ? zotero.collectionItemsRecursive(zoteroUserId, collectionKey, opts)
-      : zotero.collectionItems(zoteroUserId, collectionKey, opts);
-  });
-  h('zotero:searchItems', async (_e, library: ZoteroLibrary, query: string) => zotero.searchItems(library, query));
-  h('zotero:itemAttachments', async (_e, itemKey: string, library?: ZoteroLibrary) => {
-    const { zoteroUserId } = getSettings();
-    return zotero.itemAttachments(zoteroUserId, itemKey, library);
-  });
-
-  // works / library
-  h('home:academicSnapshot', async () => {
-    const stats = getAcademicHomeStats();
-    return {
-      stats,
-      health: getCorpusHealth(),
-      queue: scanQueue.snapshot(),
-      latestSync: getSyncLog(1)[0] ?? null,
-    };
-  });
+  // ── Corpus: works and ideas ─────────────────────────────────────────────────
   h('works:list', async (_e, filter?: WorkFilter) => works.listWorks(filter));
   h('works:listPage', async (_e, filter, request) => works.listWorksPage(filter, request));
   h('works:listZoteroTags', async () => works.listZoteroTags());
@@ -1635,21 +1260,6 @@ export function registerIpc(
     await shell.openExternal(zoteroSelectUrl(work.zotero_key));
     return { ok: true, mode: 'select' as const, page };
   });
-  h('shell:openExternal', async (_e, url: string) => {
-    // Only follow web/mail links rendered from Markdown — never arbitrary schemes.
-    if (typeof url === 'string' && /^(https?:|mailto:)/i.test(url.trim())) {
-      await shell.openExternal(url.trim());
-    }
-  });
-  h('shell:openThirdPartyNotices', async () => {
-    const noticesPath = app.isPackaged
-      ? path.join(process.resourcesPath, 'legal', 'THIRD_PARTY_NOTICES.md')
-      : path.join(app.getAppPath(), 'THIRD_PARTY_NOTICES.md');
-    if (!fs.existsSync(noticesPath)) throw new Error('No se encontraron los avisos de terceros de Nodus.');
-    const error = await shell.openPath(noticesPath);
-    if (error) throw new Error(error);
-  });
-  h('shell:openPrivacyPolicy', async () => openPrivacyPolicy());
   h('study:knowledge:processing:resolve', async (event, requestId: string, decision) => {
     resolveStudyMaterialAiProcessingRequest(event.sender.id, requestId, decision);
   });
